refactor: Minor refactors on types and wrapping
